requirements.txt
cement==2.10.2
tweepy==3.6.0
dataset
datafreeze
pytz==2018.3

google-api-python-client==1.6.5
google-auth==1.4.1
google-auth-httplib2==0.0.3
google-cloud-bigquery==0.31.0
google-auth-oauthlib==0.2.0
